:sparkles: Add answer seeding

+++ database/migrations/2019_05_09_114406_create_posts_table.php
@@ -17,8 +17,8 @@ class CreatePostsTable extends Migration
             $table->bigIncrements('id');
             $table->timestamps();
             $table->string('title');
-            $table->unsignedInteger('score');
-            $table->unsignedInteger('viewCount');
+            $table->unsignedBigInteger('score');
+            $table->unsignedBigInteger('viewCount');
             $table->longText('body');
             $table->json('tags')->default("[]");
         });
